Fixed a small bug that caused the install screen to show after conversion to JSON data files. I also decided to move the sidebars into a component as it allowed me to move a lot of the logic into a single location and make the index file a little smaller and easier to read.
