+++ index.php
@@ -6,10 +6,6 @@ require_once('common.php');
 $context_menu = file_get_contents(COMPONENTS . "/contextmenu/context_menu.json");
 $context_menu = json_decode($context_menu, true);
 
-// Right Bar
-$right_bar = file_get_contents(COMPONENTS . "/right_bar.json");
-$right_bar = json_decode($right_bar, true);
-
 // Read Components, Plugins, Themes
 $components = Common::readDirectory(COMPONENTS);
 $plugins = Common::readDirectory(PLUGINS);
@@ -44,26 +40,6 @@ if (isset($_SESSION['theme'])) {
 	<meta name="msapplication-config" content="favicons/browserconfig.xml">
 	<meta name="theme-color" content="#ffffff">
 
-	<?php
-	// Load System CSS Files
-	echo('<link rel="stylesheet" href="themes/' . $theme . '/main.min.css">');
-
-	require_once('public/minify-core-js.php');
-
-	// Load Plugin CSS Files
-	foreach ($plugins as $plugin) {
-		if (file_exists(PLUGINS . "/" . $plugin . "/screen.css")) {
-			echo('<link rel="stylesheet" href="plugins/'.$plugin.'/screen.css">');
-		}
-	}
-
-	?>
-
-
-
-</head>
-
-<body>
 	<script>
 		var i18n = (function(lang) {
 			return function(word, args) {
@@ -76,6 +52,26 @@ if (isset($_SESSION['theme'])) {
 			}
 		})(<?php echo json_encode($lang); ?>)
 	</script>
+
+	<?php
+	// Load System CSS Files
+	echo('<link rel="stylesheet" href="themes/' . $theme . '/main.min.css">');
+
+
+	require_once('public/minify-core-js.php');
+
+	// Load Plugin CSS Files
+	foreach ($plugins as $plugin) {
+		if (file_exists(PLUGINS . "/" . $plugin . "/screen.css")) {
+			echo('<link rel="stylesheet" href="plugins/'.$plugin.'/screen.css">');
+		}
+	}
+
+	?>
+
+</head>
+
+<body>
 
 	<?php
 
@@ -105,97 +101,9 @@ if (isset($_SESSION['theme'])) {
 		?>
 
 		<div id="workspace">
+		<?php require_once('components/sidebars/sb-left.php'); ?>
 
-			<div id="sb-left" class="sidebar">
-				<div id="sb-left-title">
-					<i id="sb-left-lock" class="fas fa-lock"></i>
-					<?php if (!common::isWINOS()) {
-						?>
-						<i id="finder-quick" class="fas fa-search"></i>
-						<i id="tree-search" class="fas fa-search"></i>
-						<h2 id="finder-label"> <?php i18n("Explore"); ?> </h2>
-						<div id="finder-wrapper">
-							<i id="finder-options" class="fas fa-cog"></i>
-							<div id="finder-inner-wrapper">
-								<input type="text" id="finder"></input>
-						</div>
-						<ul id="finder-options-menu" class="options-menu">
-							<li class="chosen"><a data-option="left_prefix"><?php i18n("Prefix"); ?></a></li>
-							<li><a data-option="substring"><?php i18n("Substring"); ?></a></li>
-							<li><a data-option="regexp"><?php i18n("Regular expression"); ?></a></li>
-							<li><a data-action="search"><?php i18n("Search File Contents"); ?></a></li>
-						</ul>
-					</div>
-					<?php
-				} ?>
-			</div>
 
-			<div class="sb-left-content">
-				<div id="context-menu" data-path="" data-type="">
-
-					<?php
-
-					////////////////////////////////////////////////////////////
-					// Load Context Menu
-					////////////////////////////////////////////////////////////
-
-					foreach ($context_menu as $menu_item => $data) {
-						if ($data['title'] == 'Break') {
-							echo('<hr class="'.$data['applies-to'].'">');
-						} else {
-							echo('<a class="'.$data['applies-to'].'" onclick="'.$data['onclick'].'"><i class="'.$data['icon'].'"></i>'.get_i18n($data['title']).'</a>');
-						}
-					}
-
-					foreach ($plugins as $plugin) {
-						if (file_exists(PLUGINS . "/" . $plugin . "/plugin.json")) {
-							$pdata = file_get_contents(PLUGINS . "/" . $plugin . "/plugin.json");
-							$pdata = json_decode($pdata, true);
-							if (isset($pdata[0]['contextmenu'])) {
-								foreach ($pdata[0]['contextmenu'] as $contextmenu) {
-									if ((!isset($contextmenu['admin']) || ($contextmenu['admin']) && checkAccess()) || !$contextmenu['admin']) {
-										if (isset($contextmenu['applies-to']) && isset($contextmenu['action']) && isset($contextmenu['icon']) && isset($contextmenu['title'])) {
-											echo('<hr class="'.$contextmenu['applies-to'].'">');
-											echo('<a class="'.$contextmenu['applies-to'].'" onclick="'.$contextmenu['action'].'"><i class="'.$contextmenu['icon'].'"></i>'.$contextmenu['title'].'</a>');
-										}
-									}
-								}
-							}
-						}
-					} ?>
-
-				</div>
-
-				<div id="file-manager"></div>
-
-				<ul id="list-active-files"></ul>
-
-			</div>
-
-			<div id="side-projects" class="sb-left-projects">
-				<div id="project-list" class="sb-project-list">
-
-					<div class="project-list-title">
-						<h2><?php i18n("Projects"); ?></h2>
-						<a id="projects-collapse" class="fas fa-chevron-circle-down" alt="<?php i18n("Collapse"); ?>"></a>
-						<?php if (checkAccess()) {
-							?>
-							<a id="projects-manage" class="fas fa-archive"></a>
-							<a id="projects-create" class="fas fa-plus-circle" alt="<?php i18n("Create Project"); ?>"></a>
-							<?php
-						} ?>
-					</div>
-
-					<div class="sb-projects-content"></div>
-
-				</div>
-			</div>
-
-			<div id="sb-left-handle">
-				<span>|||</span>
-			</div>
-
-		</div>
 		<div id="editor-region">
 			<div id="editor-top-bar">
 				<ul id="tab-list-active-files">
@@ -256,69 +164,7 @@ if (isset($_SESSION['theme'])) {
 				<li id="merge-all"><a> <?php i18n("Merge all"); ?> </a></li>
 			</ul>
 		</div>
-
-		<div id="sb-right" class="sidebar">
-
-			<div id="sb-right-handle">
-				<span>|||</span>
-
-			</div>
-			<div id="sb-right-title">
-				<i id="sb-right-lock" class="fas fa-unlock"></i>
-			</div>
-
-			<div class="sb-right-content">
-
-				<?php
-
-				////////////////////////////////////////////////////////////
-				// Load Right Bar
-				////////////////////////////////////////////////////////////
-
-				foreach ($right_bar as $item_rb => $data) {
-					if (!isset($data['admin'])) {
-						$data['admin'] = false;
-					}
-					if ($data['title'] == 'break') {
-						if (!$data['admin'] || $data['admin'] && checkAccess()) {
-							echo("<hr>");
-						}
-					} elseif ($data['title'] != 'break' && $data['title'] != 'pluginbar' && $data['onclick'] == '') {
-						if (!$data['admin'] || $data['admin'] && checkAccess()) {
-							echo("<div class='sb-right-category'>".get_i18n($data['title'])."</div>");
-						}
-					} elseif ($data['title'] == 'pluginbar') {
-						if (!$data['admin'] || $data['admin'] && checkAccess()) {
-							foreach ($plugins as $plugin) {
-								if (file_exists(PLUGINS . "/" . $plugin . "/plugin.json")) {
-									$pdata = file_get_contents(PLUGINS . "/" . $plugin . "/plugin.json");
-									$pdata = json_decode($pdata, true);
-									if (isset($pdata[0]['rightbar'])) {
-										foreach ($pdata[0]['rightbar'] as $rightbar) {
-											if ((!isset($rightbar['admin']) || ($rightbar['admin']) && checkAccess()) || !$rightbar['admin']) {
-												if (isset($rightbar['action']) && isset($rightbar['icon']) && isset($rightbar['title'])) {
-													echo('<a onclick="'.$rightbar['action'].'"><i class="'.$rightbar['icon'].'"></i>'.get_i18n($rightbar['title']).'</a>');
-												}
-											}
-										}
-										//echo("<hr>");
-									}
-								}
-							}
-						}
-					} else {
-						if (!$data['admin'] || $data['admin'] && checkAccess()) {
-							echo('<a onclick="'.$data['onclick'].'"><i class="'.$data['icon'].'"></i>'.get_i18n($data['title']).'</a>');
-						}
-					}
-				} ?>
-
-			</div>
-
-			<!--<div class="sidebar-handle">-->
-			<!--	<span>|||</span>-->
-			<!--</div>-->
-		</div>
+		<?php require_once('components/sidebars/sb-right.php'); ?>
 
 	</div>
